Vim swap directory is now the same as the file
Having a central swap filw directory just seems wrong.  What is you are
editing a file a shared system, or a network drive, and more than one
person is editing it?    What is if you start editing it on one system,
or account, then start editing it from another.

Killing this option unless someone has a good argument against it.